requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
djangorestframework==3.11
django-rest-swagger
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.0
python-decouple==3.3
